QuickAdd: use new component on Cart + Checkout pages
