+++ apps/store/src/graphql/ProductRecommendationFragment.graphql
@@ -1,5 +1,6 @@
 fragment ProductRecommendation on Product {
   id
+  name
   displayNameFull
   displayNameShort
   pageLink
